Create Accelerometer and GPS managers

+++ app/src/main/java/com/jj/sensorcollector/domain/sensors/SensorData.kt
@@ -2,5 +2,5 @@ package com.jj.sensorcollector.domain.sensors
 
 sealed class SensorData : ISensorData {
     data class AccelerometerData(val x: Float, val y: Float, val z: Float) : SensorData()
-    data class GPSData(val lat: Long, val lng: Long) : SensorData()
+    data class GPSData(val lat: Double, val lng: Double) : SensorData()
 }
